Feedback do Medico feito

+++ app/src/main/java/com/example/projeto_ttc2/presentation/ui/components/StepsCard.kt
@@ -19,7 +19,7 @@ import com.example.projeto_ttc2.presentation.ui.theme.TealGreen
 
 @Composable
 fun StepsCard(
-    steps: Long,
+    steps: Int,
     goal: Long,
     distanceKm: Double,
     onClick: () -> Unit,
